signalfxexporter: Refactor httpSender
